refactor: move a function to lib
