+++ lib/package.json
@@ -9,6 +9,7 @@
   "dependencies": {
     "@actions/core": "1.11.1",
     "@actions/exec": "^1.1.1",
+    "@actions/github": "^6.0.0",
     "js-yaml": "4.1.0",
     "zod": "^3.22.4"
   },
